general tidying up

+++ bin/scaffold_wp_plugin.sh
@@ -8,23 +8,34 @@ version="1.0.0"
 author="Brett"
 author_uri="www.thoughtworks.com"
 
-template_dir=` dirname \`dirname $0\` `"/template_files"
-plugin_dir="`pwd`/$plugin_name"
-if [ -d "$plugin_dir" ]; then
-    echo "directory $plugin_name already exists";
-    exit 1
-fi
+function get_template_dir() {
+    if [ -l "$0" ]; then
+        echo "$(dirname $(dirname $(readlink "$0") ) )/template_files"
+    else
+        echo "$(dirname $(dirname "$0" ) )/template_files"
+    fi
+}
 
-mkdir "$plugin_dir"
+function get_plugin_dir() {
+    echo "$(pwd)/$plugin_name"
+}
 
-current_dir=`pwd`
-cd "$template_dir"
-tar -vcf "$plugin_dir/template_files.tar" *
-cd "$plugin_dir"
-tar -xvf template_files.tar
-rm -f template_files.tar
+function make_plugin_dir() {
+    if [ -d "$plugin_dir" ]; then
+        echo "directory $plugin_name already exists";
+        exit 1
+    fi
+
+    mkdir "$plugin_dir"
+}
 
 function do_subs() {
+    if [ ! -d "$plugin_dir" ]; then
+        echo "$plugin_dir doesn't exist can't rename files"
+        return 1
+    fi
+
+    cd "$plugin_dir"
     for sub_filename in `find . -name "*.php"`; do
 
         sed -i ".plugin_name.bkp" s/"{plugin_name}"/"$plugin_name"/g "$sub_filename"
@@ -35,24 +46,61 @@ function do_subs() {
         sed -i ".author_uri.bkp" s/"{author_uri}"/"$author_uri"/g "$sub_filename"
         sed -i ".plugin_uri.bkp" s/"{plugin_uri}"/"$plugin_uri"/g "$sub_filename"
     done
+
+    return 0
 }
 
 function fix_filenames() {
+    if [ ! -d "$plugin_dir" ]; then
+        echo "$plugin_dir doesn't exist can't rename files"
+        return 1
+    fi
+
+    cd "$plugin_dir"
     for plugin_file in `find . -name "*plugin_name*"`; do
-        src_filename=`echo $plugin_file | sed s/"\.\/"//g`
-        target_filename=`echo $src_filename | sed s/plugin_name/"$plugin_name"/g`
+        src_filename="$(echo $plugin_file | sed s/"\.\/"//g)"
+        target_filename="$(echo $src_filename | sed s/plugin_name/"$plugin_name"/g)"
 
         mv "$src_filename" "$target_filename"
     done
+
+    cd "$current_dir"
+    return 0
 }
 
 function remove_backup_files() {
-find . -name "*.bkp" -exec rm -f "{}" ";"
+    if [ ! -d "$plugin_dir" ]; then
+        echo "$plugin_dir doesn't exist can't rename files"
+        return 1
+    fi
+
+    cd "$plugin_dir"
+    find . -name "*.bkp" -exec rm -f "{}" ";"
+
+    return 0
 }
 
 function sed_esc() {
     echo "$1" | sed s/"\/"/"\\\\\/"/g | sed s/"&"/"\\\&"/g
 }
+
+function copy_template_files_to_plugin_dir() {
+    make_plugin_dir
+    cd "$template_dir"
+    tar -vcf "$plugin_dir/template_files.tar" *
+    cd "$plugin_dir"
+    tar -xvf template_files.tar
+    rm -f template_files.tar
+    cd "$current_dir"
+}
+
+export plugin_dir="$(get_plugin_dir)"
+export template_dir="$(get_template_dir)"
+export current_dir="$(pwd)"
+
+echo "plugin_dir = $plugin_dir"
+echo "template_dir = $template_dir"
+echo "current_dir = $current_dir"
 
 plugin_name=`sed_esc $plugin_name`
 packge_name=`sed_esc $package_name`
@@ -62,6 +110,7 @@ version=`sed_esc $version`
 author=`sed_esc $author`
 author_uri=`sed_esc $author_uri`
 
+copy_template_files_to_plugin_dir
 fix_filenames
 do_subs
